+++ nflow-engine/src/main/java/io/nflow/engine/internal/workflow/MaintenanceWorkflowStarter.java
@@ -44,8 +44,8 @@ public class MaintenanceWorkflowStarter {
     apply(env, "archive", builder::withArchiveWorkflows);
     apply(env, "delete", builder::withDeleteWorkflows);
     apply(env, "deleteArchived", builder::withDeleteArchivedWorkflows);
-    builder.withDeleteExpiredExecutorsAfter(
-        Period.parse(env.getRequiredProperty("nflow.maintenance.initial.deleteExpiredExecutorsAfter")));
+    builder.withDeleteExpiredExecutorsOlderThan(
+        Period.parse(env.getRequiredProperty("nflow.maintenance.initial.deleteExpiredExecutors.olderThan")));
     this.initialConfiguration = builder.build();
   }