+++ nflow-engine/src/main/java/io/nflow/engine/workflow/curated/MaintenanceWorkflow.java
@@ -62,8 +62,8 @@ public class MaintenanceWorkflow extends CronWorkflow {
     add(sb, "Archived", results.archivedWorkflows);
     add(sb, "Deleted", results.deletedWorkflows);
     add(sb, "Deleted archived", results.deletedArchivedWorkflows);
-    if (conf.deleteExpiredExecutorsAfter != null) {
-      int deletedExecutors = maintenanceService.cleanupExecutors(conf.deleteExpiredExecutorsAfter);
+    if (conf.deleteExpiredExecutorsOlderThan != null) {
+      int deletedExecutors = maintenanceService.cleanupExecutors(conf.deleteExpiredExecutorsOlderThan);
       add(sb, "Deleted executors", deletedExecutors);
     }
     if (sb.length() == 0) {